Udated MPU6050 - calibrated and interrupt handling , claibrated adc, Change the BH1750 to Highres Single shot

+++ test_lms_limits/src/SoilMoisture/SoilMoisture.h
@@ -10,8 +10,8 @@ private:
     uint8_t pin;
     SoilMoistureData lastReading;
     bool isInitialized;
-    float calibrationOffsetMin;  // ADC value at dry (0%)
-    float calibrationOffsetMax;  // ADC value at wet (100%)
+    uint32_t calibrationOffsetMin;  // ADC value at dry (0%)
+    uint32_t calibrationOffsetMax;  // ADC value at wet (100%)
 
 public:
     SoilMoistureSensor(uint8_t sensorPin = SOIL_MOISTURE_PIN);
